Did lots of work on making the various classes more "pluggable".
 - The major classes should now implement a valid_params() method that
returns a hashref of what parameters they accept in their new()
methods.  The format of the hash is an extension of the
Params::Validate format, with an additional 'parse' entry that
describes how to parse the entry from a $r->dir_config value.  This
greatly simplifies argument processing (anyone care to find the typo
that can't happen anymore? =).

 - HTML::Mason::ApacheHandler is now a method handler.  It's
configured the same way it's always been documented, but now it's
subclassable (and could be a subclass itself if we want to reorganize).

 - HTML::Mason::ApacheHandler no longer requires Apache::Request - CGI
can be used again, but Apache::Request is still the default.

 - The ApacheHandler 'top_level_predicate' is always defined now -
default is sub () {1} .

The HTML::Mason::Tests module will now show the entire output when
'actual' and 'expect' differ.  The "3-line neighborhood" usually
didn't include the section one was really interested in.

Now requires Params::Validate version 0.07.


git-svn-id: file:///home/autarch/tmp/Mason/trunk@1208 878abba1-507e-414f-af30-0bba99d2f2a9