make tests more likely to run: if 10 seconds pass waiting for the httpd.pid file to dissapear, we try to kill the process again.  If we get a 'no such (file|proc)' error, we remove the httpd.pid file manually and go on.  On my box, this has been quite helpful.
Also changed the regex against $! in the same section to check for the
error message 'No such file or directory' which is what I was getting
when trying to kill a non-existent process.

fix HEAD test.  For some reason, Mason was actually processing the
body of the component for a head request.  This shouldn't happen.
This test has been fixed to use headers only and to not expect the
body to be processed.  This basically catches two errors at once (arg
handling for HEAD requests, and whether a component body is executed).


git-svn-id: file:///home/autarch/tmp/Mason/trunk@1374 878abba1-507e-414f-af30-0bba99d2f2a9