the default document root bits	should be done in the resolver itself, I	think.
This removes some code	elsewhere and more importantly,	makes the
default work with or without a handler.pl file.

It's a	special	case but	I think	in this	case it	is worth it for	the
high convenience factor.  In the future, when the ApacheHandler	and
CGIHandler classes become Interp subclasses (or	something along	those
architectural lines) this hack	can go away.


git-svn-id: file:///home/autarch/tmp/Mason/trunk@1632 878abba1-507e-414f-af30-0bba99d2f2a9