fixes filter bugs by ...
defining a scalar $__filter__ which is part of the component header,
and is basically equivalent to a lexically scoped variable declared in
a <%once> section

then $self->{filter} (where $self is the component object) contains a
_reference_ to this scalar.  The scalar is actually populated when the
component is _executed_ (as part of the 'main' dynamic sub when
<%shared> is in effect).

This means that $__filter__ is a closure defined in the same scope as
%ARGS & and variables declared in <%args>, _but_ it cannot be broken
by a "return" in the main code body, because it is defined before
user-defined "main" code is executed.

Whee, it works!


git-svn-id: file:///home/autarch/tmp/Mason/trunk@2685 878abba1-507e-414f-af30-0bba99d2f2a9