Reverted back to Ken's version with an extra comment.
Detailed information from Ken:
"It uses foreach because join('', ...) uses more memory.  I turned it
into foreach when I was working on reducing memory requirements for the
Really Large Components, but it affects all components of any size.
join('', ...) uses more memory because it must allocate a temporary SV
to store the result of the join, then append that SV's contents to the
buffer.  The foreach version doesn't need any temporary internal
variables."


git-svn-id: file:///home/autarch/tmp/Mason/trunk@3067 878abba1-507e-414f-af30-0bba99d2f2a9