Fix precedence error; "and" is lower precedence than "return"
Perl 5.19.4 and above produce a warning to detect this common error; it
rightly detects that use_dhandlers and use_autohandlers incorrectly
returned true when they were set to the empty string.  The behavior in
both cases was _equivalent_ to disabling the feature, because the empty
filename could never be found to be used.

Fix the precedence using explicit parens, and add explicit tests for the
return values of use_dhandlers and use_autohandlers.  This fixes
[rt.cpan.org #87050].